chore: rename auth routes
