+++ app/routes.ts
@@ -3,8 +3,8 @@ import { index, layout, route } from "@react-router/dev/routes";
 
 export default [
   layout("./routes/_auth.tsx", [
-    route("login", "./routes/_auth.login.tsx"),
-    route("register", "./routes/_auth.register.tsx"),
+    route("signin", "./routes/_auth.signin.tsx"),
+    route("signup", "./routes/_auth.signup.tsx"),
   ]),
   layout("components/shell.tsx", [index("./routes/_dashboard._index.tsx")]),
   // ...prefix("api", [
